Handle 404 NOT FOUND in login
